refactor(siteNav): I guess now I can't camelCase? ¯\_(ツ)_/¯
